+++ .idea/encodings.xml
@@ -5,6 +5,7 @@
     <file url="file://$PROJECT_DIR$/miao-learning-base/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/miao-learning-content/miao-learning-content-api/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/miao-learning-content/miao-learning-content-api/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/miao-learning-content/miao-learning-content-api/src/main/resources/application.yml" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/miao-learning-content/miao-learning-content-model/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/miao-learning-content/miao-learning-content-model/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/miao-learning-content/miao-learning-content-service/src/main/java" charset="UTF-8" />
