+++ miao-learning-base/src/main/java/com/miao/base/model/PageResult.java
@@ -1,0 +1,37 @@
+package com.miao.base.model;
+
+import lombok.Data;
+import lombok.ToString;
+
+import java.io.Serializable;
+import java.util.List;
+
+/**
+ * @author miao
+ * @version 1.0.0
+ * @description 分页查询结果模型类
+ * @title PageResult
+ * @date 2023/1/18 17:08
+ */
+@Data
+@ToString
+public class PageResult<T> implements Serializable {
+    // 数据列表
+    private List<T> items;
+
+    //总记录数
+    private long counts;
+
+    //当前页码
+    private long page;
+
+    //每页记录数
+    private long pageSize;
+
+    public PageResult(List<T> items, long counts, long page, long pageSize) {
+        this.items = items;
+        this.counts = counts;
+        this.page = page;
+        this.pageSize = pageSize;
+    }
+}
