+++ miao-learning-content/miao-learning-content-api/src/main/java/com/miao/content/api/CourseBaseInfoController.java
@@ -1,0 +1,39 @@
+package com.miao.content.api;
+
+import com.miao.base.model.PageParams;
+import com.miao.base.model.PageResult;
+import com.miao.content.model.dto.QueryCourseParamsDto;
+import com.miao.content.model.po.CourseBase;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @author miao
+ * @version 1.0.0
+ * @description TODO
+ * @title CourseBaseInfoController
+ * @date 2023/1/18 17:19
+ */
+@Api(value = "课程信息编辑接口",tags = "课程信息编辑接口")
+@RestController
+public class CourseBaseInfoController {
+
+    @ApiOperation("课程查询接口")
+    @PostMapping("/course/list")
+    public PageResult<CourseBase> list(PageParams pageParams, @RequestBody QueryCourseParamsDto queryCourseParamsDto){
+        CourseBase courseBase = new CourseBase();
+        courseBase.setName("测试名称");
+        courseBase.setCreateDate(LocalDateTime.now());
+        List<CourseBase> courseBases = new ArrayList();
+        courseBases.add(courseBase);
+        PageResult pageResult = new PageResult<CourseBase>(courseBases,10,1,10);
+        return pageResult;
+    }
+}
