+++ miao-learning-content/miao-learning-content-api/src/main/resources/application.yml
@@ -1,0 +1,19 @@
+server:
+  servlet:
+    context-path: /content
+  port: 63040
+
+# ?????
+spring:
+  application:
+    name: content-api
+logging:
+  config: classpath:log4j2-dev.xml
+
+
+swagger:
+  title: "喵喵秒"
+  description: "内容系统管理系统对课程相关信息进行业务管理数据"
+  base-package: com.miao.content
+  enabled: true
+  version: 1.0.0
