--- miao-learning-content/miao-learning-content-model/src/main/java/com/miao/content/model/MiaoLearningContentModelApplication.java
@@ -1,13 +1,0 @@
-package com.miao.content.model;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MiaoLearningContentModelApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MiaoLearningContentModelApplication.class, args);
-	}
-
-}
