--- miao-learning-generator/src/main/java/com/miao/content/controller/TeachplanController.java
@@ -1,25 +1,0 @@
-package com.miao.content.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import com.miao.content.service.TeachplanService;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-
-/**
- * <p>
- * 课程计划 前端控制器
- * </p>
- *
- * @author itcast
- */
-@Slf4j
-@RestController
-@RequestMapping("teachplan")
-public class TeachplanController {
-
-    @Autowired
-    private TeachplanService  teachplanService;
-}
