--- miao-learning-generator/src/main/java/com/miao/content/controller/TeachplanMediaController.java
@@ -1,25 +1,0 @@
-package com.miao.content.controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import com.miao.content.service.TeachplanMediaService;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-
-/**
- * <p>
- *  前端控制器
- * </p>
- *
- * @author itcast
- */
-@Slf4j
-@RestController
-@RequestMapping("teachplanMedia")
-public class TeachplanMediaController {
-
-    @Autowired
-    private TeachplanMediaService  teachplanMediaService;
-}
