--- miao-learning-generator/src/main/java/com/miao/content/mapper/CourseBaseMapper.java
@@ -1,16 +1,0 @@
-package com.miao.content.mapper;
-
-import com.miao.content.model.po.CourseBase;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- * 课程基本信息 Mapper 接口
- * </p>
- *
- * @author itcast
- */
-public interface CourseBaseMapper extends BaseMapper<CourseBase> {
-
-}
