--- miao-learning-generator/src/main/java/com/miao/content/mapper/CourseCategoryMapper.java
@@ -1,16 +1,0 @@
-package com.miao.content.mapper;
-
-import com.miao.content.model.po.CourseCategory;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- * 课程分类 Mapper 接口
- * </p>
- *
- * @author itcast
- */
-public interface CourseCategoryMapper extends BaseMapper<CourseCategory> {
-
-}
