--- miao-learning-generator/src/main/java/com/miao/content/mapper/CourseMarketMapper.java
@@ -1,16 +1,0 @@
-package com.miao.content.mapper;
-
-import com.miao.content.model.po.CourseMarket;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- * 课程营销信息 Mapper 接口
- * </p>
- *
- * @author itcast
- */
-public interface CourseMarketMapper extends BaseMapper<CourseMarket> {
-
-}
