--- miao-learning-generator/src/main/java/com/miao/content/mapper/CourseTeacherMapper.java
@@ -1,16 +1,0 @@
-package com.miao.content.mapper;
-
-import com.miao.content.model.po.CourseTeacher;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- * 课程-教师关系表 Mapper 接口
- * </p>
- *
- * @author itcast
- */
-public interface CourseTeacherMapper extends BaseMapper<CourseTeacher> {
-
-}
