--- miao-learning-generator/src/main/java/com/miao/content/mapper/CourseTeacherMapper.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.miao.content.mapper.CourseTeacherMapper">
-
-    <!-- 通用查询映射结果 -->
-    <resultMap id="BaseResultMap" type="com.miao.content.model.po.CourseTeacher">
-        <id column="id" property="id" />
-        <result column="course_id" property="courseId" />
-        <result column="teacher_name" property="teacherName" />
-        <result column="position" property="position" />
-        <result column="introduction" property="introduction" />
-        <result column="photograph" property="photograph" />
-        <result column="create_date" property="createDate" />
-    </resultMap>
-
-    <!-- 通用查询结果列 -->
-    <sql id="Base_Column_List">
-        id, course_id, teacher_name, position, introduction, photograph, create_date
-    </sql>
-
-</mapper>
