--- miao-learning-generator/src/main/java/com/miao/content/mapper/TeachplanMapper.java
@@ -1,16 +1,0 @@
-package com.miao.content.mapper;
-
-import com.miao.content.model.po.Teachplan;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- * 课程计划 Mapper 接口
- * </p>
- *
- * @author itcast
- */
-public interface TeachplanMapper extends BaseMapper<Teachplan> {
-
-}
