--- miao-learning-generator/src/main/java/com/miao/content/mapper/TeachplanMediaMapper.java
@@ -1,16 +1,0 @@
-package com.miao.content.mapper;
-
-import com.miao.content.model.po.TeachplanMedia;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
-
-/**
- * <p>
- *  Mapper 接口
- * </p>
- *
- * @author itcast
- */
-public interface TeachplanMediaMapper extends BaseMapper<TeachplanMedia> {
-
-}
