--- miao-learning-generator/src/main/java/com/miao/content/mapper/TeachplanMediaMapper.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.miao.content.mapper.TeachplanMediaMapper">
-
-    <!-- 通用查询映射结果 -->
-    <resultMap id="BaseResultMap" type="com.miao.content.model.po.TeachplanMedia">
-        <id column="id" property="id" />
-        <result column="media_id" property="mediaId" />
-        <result column="teachplan_id" property="teachplanId" />
-        <result column="course_id" property="courseId" />
-        <result column="media_fileName" property="mediaFilename" />
-        <result column="create_date" property="createDate" />
-        <result column="create_people" property="createPeople" />
-        <result column="change_people" property="changePeople" />
-    </resultMap>
-
-    <!-- 通用查询结果列 -->
-    <sql id="Base_Column_List">
-        id, media_id, teachplan_id, course_id, media_fileName, create_date, create_people, change_people
-    </sql>
-
-</mapper>
