--- miao-learning-generator/src/main/java/com/miao/content/model/po/CourseTeacher.java
@@ -1,63 +1,0 @@
-package com.miao.content.model.po;
-
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
-import java.time.LocalDateTime;
-import com.baomidou.mybatisplus.annotation.FieldFill;
-import com.baomidou.mybatisplus.annotation.TableField;
-import java.io.Serializable;
-import lombok.Data;
-import com.baomidou.mybatisplus.annotation.TableName;
-
-/**
- * <p>
- * 课程-教师关系表
- * </p>
- *
- * @author itcast
- */
-@Data
-@TableName("course_teacher")
-public class CourseTeacher implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * 主键
-     */
-    @TableId(value = "id", type = IdType.AUTO)
-    private Long id;
-
-    /**
-     * 课程标识
-     */
-    private Long courseId;
-
-    /**
-     * 教师标识
-     */
-    private String teacherName;
-
-    /**
-     * 教师职位
-     */
-    private String position;
-
-    /**
-     * 教师简介
-     */
-    private String introduction;
-
-    /**
-     * 照片
-     */
-    private String photograph;
-
-    /**
-     * 创建时间
-     */
-    @TableField(fill = FieldFill.INSERT)
-    private LocalDateTime createDate;
-
-
-}
