--- miao-learning-generator/src/main/java/com/miao/content/service/CourseMarketService.java
@@ -1,16 +1,0 @@
-package com.miao.content.service;
-
-import com.miao.content.model.po.CourseMarket;
-import com.baomidou.mybatisplus.extension.service.IService;
-
-/**
- * <p>
- * 课程营销信息 服务类
- * </p>
- *
- * @author itcast
- * @since 2022-12-10
- */
-public interface CourseMarketService extends IService<CourseMarket> {
-
-}
