--- miao-learning-generator/src/main/java/com/miao/content/service/TeachplanMediaService.java
@@ -1,16 +1,0 @@
-package com.miao.content.service;
-
-import com.miao.content.model.po.TeachplanMedia;
-import com.baomidou.mybatisplus.extension.service.IService;
-
-/**
- * <p>
- *  服务类
- * </p>
- *
- * @author itcast
- * @since 2022-12-10
- */
-public interface TeachplanMediaService extends IService<TeachplanMedia> {
-
-}
