--- miao-learning-generator/src/main/java/com/miao/content/service/TeachplanService.java
@@ -1,16 +1,0 @@
-package com.miao.content.service;
-
-import com.miao.content.model.po.Teachplan;
-import com.baomidou.mybatisplus.extension.service.IService;
-
-/**
- * <p>
- * 课程计划 服务类
- * </p>
- *
- * @author itcast
- * @since 2022-12-10
- */
-public interface TeachplanService extends IService<Teachplan> {
-
-}
