--- miao-learning-generator/src/main/java/com/miao/content/service/impl/CourseCategoryServiceImpl.java
@@ -1,23 +1,0 @@
-package com.miao.content.service.impl;
-
-import com.miao.content.model.po.CourseCategory;
-import com.miao.content.mapper.CourseCategoryMapper;
-import com.miao.content.service.CourseCategoryService;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
-
-/**
- * <p>
- * 课程分类 服务实现类
- * </p>
- *
- * @author itcast
- */
-@Slf4j
-@Service
-public class CourseCategoryServiceImpl extends ServiceImpl<CourseCategoryMapper, CourseCategory> implements CourseCategoryService {
-
-}
