--- miao-learning-generator/src/main/java/com/miao/content/service/impl/TeachplanMediaServiceImpl.java
@@ -1,23 +1,0 @@
-package com.miao.content.service.impl;
-
-import com.miao.content.model.po.TeachplanMedia;
-import com.miao.content.mapper.TeachplanMediaMapper;
-import com.miao.content.service.TeachplanMediaService;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
-
-/**
- * <p>
- *  服务实现类
- * </p>
- *
- * @author itcast
- */
-@Slf4j
-@Service
-public class TeachplanMediaServiceImpl extends ServiceImpl<TeachplanMediaMapper, TeachplanMedia> implements TeachplanMediaService {
-
-}
