+++ miao-learning-generator/src/main/java/com/miao/generator/BankCodeGenerator.java
@@ -36,8 +36,8 @@ public class BankCodeGenerator {
 		// 全局配置
 		GlobalConfig gc = new GlobalConfig();
 		gc.setFileOverride(true);
-		gc.setOutputDir(System.getProperty("user.dir") + "/xuecheng-plus-generator/src/main/java");
-		gc.setAuthor("itcast");
+		gc.setOutputDir(System.getProperty("user.dir") + "/miao-learning-generator/src/main/java");
+		gc.setAuthor("miao");
 		gc.setOpen(false);
 		gc.setSwagger2(false);
 		gc.setServiceName("%sService");
